# docs/api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: AI Enhanced PDF Scholar API
  description: |
    **Enterprise-grade intelligent document management platform for academic research**
    
    This API provides comprehensive functionality for:
    - 🎯 **Intelligent Document Processing**: Smart import with metadata extraction and duplicate detection
    - 📊 **Advanced Citation Analysis**: Multi-format parsing with network discovery
    - 🧠 **RAG (Retrieval-Augmented Generation)**: AI-powered document queries and analysis
    - 🔐 **Enterprise Security**: JWT authentication with RBAC and XSS protection
    - ⚡ **Real-time Features**: WebSocket support for streaming operations
    - 📈 **Performance Monitoring**: APM with metrics and health checks
    
    ## Authentication
    
    Most endpoints require JWT authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    
    API endpoints are rate-limited to ensure fair usage:
    - **Default**: 100 requests per minute per IP
    - **Global IP**: 1000 requests per hour per IP
    - **Authenticated users**: Higher limits apply
    
    ## Error Handling
    
    The API uses standard HTTP status codes and provides detailed error information:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found (resource not found)
...
    
  version: "2.1.0"
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/documents:
    get:
      tags: [documents]
      summary: List documents
      description: Get paginated list of documents with filtering options
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of documents per page
        - name: search_query
          in: query
          schema:
            type: string
          description: Search term to filter documents
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [title, created_at, updated_at, file_size, file_path]
            default: created_at
          description: Field to sort by
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
        - name: show_missing
          in: query
          schema:
            type: boolean
            default: false
          description: Include documents with missing files
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [documents]
      summary: Upload document
      description: Upload and process a new document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF, DOCX, or TXT file to upload
                title:
                  type: string
                  description: Document title (auto-extracted if not provided)
                check_duplicates:
                  type: boolean
                  default: true
                  description: Check for duplicate documents
                extract_metadata:
                  type: boolean
                  default: true
                  description: Extract metadata from document
                tags:
                  type: array
                  items:
                    type: string
                  description: Tags to associate with document
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentImportResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Duplicate document detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '413':
          description: File too large
...
  /api/rag/query:
    post:
      tags: [rag]
      summary: Query documents with AI
      description: |
        Query documents using RAG (Retrieval-Augmented Generation) for intelligent answers.
        Supports context-aware queries with semantic search and citation tracking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Natural language query
                  example: "What are the main findings about transformer architectures?"
                document_id:
                  type: integer
                  description: Specific document to query (optional)
                context_window:
                  type: integer
                  default: 4000
                  description: Context window size for query processing
                include_sources:
                  type: boolean
                  default: true
                  description: Include source citations in response
                temperature:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 0.1
                  description: Response creativity (0=factual, 1=creative)
      responses:
        '200':
          description: RAG query completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGQueryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Document not found or not indexed
...
  /ws/{client_id}:
    get:
      tags: [system]
      summary: WebSocket connection (Not REST)
      description: |
        **WebSocket Endpoint - Not a REST API**
        
        Connect to real-time WebSocket for streaming operations:
        - RAG query progress updates
        - Document processing status
        - System notifications
        
        **Connection URL**: `ws://localhost:8000/ws/{client_id}`
        
        **Message Types**:
        - `ping/pong` - Keep alive
        - `rag_query` - Submit RAG query
        - `rag_progress` - Query progress update
        - `rag_response` - Query result
        - `rag_error` - Query error
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
          description: Unique client identifier
      responses:
        '101':
          description: WebSocket connection established
        '400':
          description: Invalid client ID
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
  schemas:
...
    BaseResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Human-readable message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error_code:
              type: string
              description: Machine-readable error code
            details:
              type: object
              description: Additional error details
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            validation_errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
    ConflictError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            existing_resource:
              type: object
              description: Information about existing conflicting resource

    # Authentication Models
...
    DocumentListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                documents:
                  type: array
                  items:
                    $ref: '#/components/schemas/Document'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    DocumentImportResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/Document'
                - type: object
                  properties:
                    processing_stats:
                      type: object
                      properties:
                        processing_time:
                          type: number
                          description: Processing time in seconds
                        extracted_metadata:
                          type: boolean
                        indexed_for_rag:
                          type: boolean
    Document:
      type: object
      properties:
        id:
          type: integer
          description: Unique document identifier
        title:
          type: string
          description: Document title
        file_path:
          type: string
          description: File path (relative to upload directory)
        file_size:
          type: integer
          description: File size in bytes
        content_hash:
          type: string
          description: SHA-256 hash of file content
        mime_type:
          type: string
          description: MIME type of the file
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_file_available:
          type: boolean
          description: Whether the physical file exists
        metadata:
          type: object
          properties:
            author:
              type: string
            subject:
              type: string
            keywords:
              type: array
              items:
                type: string
            page_count:
              type: integer
            word_count:
              type: integer
        tags:
          type: array
          items:
            type: string
        rag_indexed:
          type: boolean
          description: Whether document is indexed for RAG queries
...
    RAGQueryResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                query:
                  type: string
                  description: Original query
                answer:
                  type: string
                  description: AI-generated answer
                sources:
                  type: array
                  items:
                    type: object
                    properties:
                      document_id:
                        type: integer
                      document_title:
                        type: string
                      page_number:
                        type: integer
                      confidence:
                        type: number
                      excerpt:
                        type: string
                processing_time:
                  type: number
                  description: Query processing time in seconds
                context_tokens:
                  type: integer
                  description: Number of tokens used for context
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean